test: regroup added tests into a few thematic files by function
The hardening work scattered tests across many small single-purpose
files. Regroup everything I added into a few thematic files, leaving the
source-shaped test files (conv/map/slice/struct/interface/value/call/
func/extern/util) untouched. No test logic changes (210 -> 210 funcs,
suite green). convert test files 21 -> 15.

Black-box (convert_test):
- keys_test:       map keys + deterministic ordering
                   (hashable + order + identity + decorate)
- conversion_test: value conversion correctness/safety
                   (checked + unwrap + review_misc: Duration/typed-nil/...)
- robustness_test: no-panic + bridge contracts + freeze/concurrency
                   (panics + safestring + bridge_contract + freeze_race)
White-box (convert): internal_test (already consolidated)
Root: starlight_features_test (dialect + WithGlobals); starlight_test untouched

Each file opens with a commented section list. Also documents the test
organization rule in CLAUDE.md so future tests follow it.
